So i made a pathfinding thing for a game not for real life use!
i added a map overlay in  pathconvergance.html?key= where the key is your google map key m files are to run a simulation and capture photos and segnet is well understood 13.py if inferencing segnet files. send the zip file to main branch and deleted the zip file georef takes osm data and pumps it out with a special file name! news boundaries are in the name are are used by pathconvergance.html